feat(bin): add cli
