+++ src/main/Deferred.ts
@@ -1,0 +1,20 @@
+export class Deferred<T extends any> {
+    promise: Promise<T>;
+    private _resolve!: (value?: T) => void;
+    private _reject!: (reason?: Error) => void;
+
+    constructor() {
+        this.promise = new Promise((resolve, reject) => {
+            this._resolve = resolve;
+            this._reject = reject;
+        });
+    }
+
+    resolve(value?: any) {
+        this._resolve(value);
+    }
+
+    reject(reason?: Error) {
+        this._reject(reason);
+    }
+}
