+++ src/renderer/index.css
@@ -4,8 +4,12 @@ body,
     width: 100%;
     height: 100%;
     box-sizing: border-box;
+    padding: 0.2rem 0.1rem;
     margin: 0;
-    padding: 0;
+}
+
+#app {
+    margin: 0;
 }
 
 #app .CodeMirror {
